rename: 컴포넌트 디렉토리 임시 통합
