+++ src/components/CharactorSelect.tsx
