+++ src/components/index.ts
@@ -1,0 +1,7 @@
+export { default as CharactorSelect } from "./CharactorSelect";
+export { default as CheckBox } from "./CheckBox";
+export { default as CheckList } from "./CheckList";
+export { default as ColorSelect } from "./ColorSelect";
+export { default as CustomCharacter } from "./CustomCharacter";
+export { default as HeaderLayout } from "./HeaderLayout";
+export { default as NicknameInput } from "./NicknameInput";
